+++ packages/input.number/src/NumberInput.tsx
@@ -4,6 +4,7 @@ import {IcRoundNumbers} from '@alinea/ui/icons/IcRoundNumbers'
 import {NumberField} from './NumberField'
 import css from './NumberInput.module.scss'
 import {fromModule} from '@alinea/ui'
+import {useState} from 'react'
 
 const styles = fromModule(css)
 
@@ -23,7 +24,8 @@ export function NumberInput({state, field}: NumberInputProps) {
     maxValue,
     step
   } = field.options
-  const [value = initialValue, setValue] = useInput(state)
+  const [value, setValue] = useInput(state)
+  const [defined, setDefined] = useState<boolean>(typeof value === undefined)
 
   return (
     <InputLabel
@@ -38,14 +40,25 @@ export function NumberInput({state, field}: NumberInputProps) {
       <input
         type="number"
         className={styles.root.input()}
-        value={String(value ?? '')}
-        onChange={e =>
+        value={
+          String(value ?? '') || (!defined ? String(initialValue ?? '') : '')
+        }
+        onChange={e => {
           setValue(
             e.currentTarget.value === ''
               ? undefined!
               : Number(e.currentTarget.value)
           )
-        }
+          if (!defined) setDefined(true)
+        }}
+        onBlur={e => {
+          if (
+            !optional &&
+            value === undefined &&
+            typeof initialValue !== 'undefined'
+          )
+            setValue(initialValue)
+        }}
         min={minValue}
         max={maxValue}
         step={step || 1}